+++ backend/src/main/java/pw/checkers/data/GameState.java
@@ -1,4 +1,4 @@
-package pw.checkers.pojo;
+package pw.checkers.data;
 
 
 import lombok.AllArgsConstructor;
