+++ backend/src/main/java/pw/checkers/data/enums/PieceType.java
@@ -1,4 +1,4 @@
-package pw.checkers.pojo;
+package pw.checkers.data.enums;
 
 public enum PieceType {
     PAWN, KING
