+++ backend/src/main/java/pw/checkers/messages/JoinMessage.java
@@ -1,0 +1,16 @@
+package pw.checkers.messages;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class JoinMessage {
+    private String gameId;
+    private String color;
+    private User opponent;
+}
