+++ backend/src/main/java/pw/checkers/messages/Message.java
@@ -1,4 +1,4 @@
-package pw.checkers.pojo;
+package pw.checkers.messages;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,8 +9,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class JoinMessage {
-    private String gameId;
-    private String color;
-    private User opponent;
+public class Message<T> {
+    private String type;
+    private T content;
 }
