+++ backend/src/main/java/pw/checkers/messages/PossibilitiesInput.java
@@ -1,4 +1,4 @@
-package pw.checkers.pojo;
+package pw.checkers.messages;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,8 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Message<T> {
-    private String type;
-    private T content;
+public class PossibilitiesInput {
+    private String gameId;
+    private int row;
+    private int col;
 }
