+++ backend/src/main/java/pw/checkers/messages/PromptMessage.java
@@ -1,4 +1,4 @@
-package pw.checkers.pojo;
+package pw.checkers.messages;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 
 @Getter
 @Setter
-@AllArgsConstructor
 @NoArgsConstructor
-public class GameEnd {
-    private String result;
+@AllArgsConstructor
+public class PromptMessage {
+    private String message;
 }
