--- backend/src/main/java/pw/checkers/pojo/User.java
@@ -1,14 +1,0 @@
-package pw.checkers.pojo;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class User {
-    private String username;
-}
