+++ backend/src/main/java/pw/checkers/service/GameService.java
@@ -1,9 +1,9 @@
 package pw.checkers.service;
 
-import pw.checkers.pojo.GameState;
-import pw.checkers.pojo.Move;
-import pw.checkers.pojo.MoveOutput;
-import pw.checkers.pojo.PossibleMoves;
+import pw.checkers.data.GameState;
+import pw.checkers.messages.Move;
+import pw.checkers.messages.MoveOutput;
+import pw.checkers.messages.PossibleMoves;
 
 public interface GameService {
     GameState createGame();
