+++ backend/src/main/java/pw/checkers/sockets/CheckersWebSocketHandler.java
@@ -7,7 +7,8 @@ import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.TextMessage;
 import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.TextWebSocketHandler;
-import pw.checkers.pojo.*;
+import pw.checkers.data.GameState;
+import pw.checkers.messages.*;
 import pw.checkers.service.GameService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
