+++ backend/src/main/java/pw/checkers/web/GameController.java
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pw.checkers.pojo.GameState;
+import pw.checkers.data.GameState;
 import pw.checkers.service.GameService;
 
 @RestController
